+++ gstore/logger.py
@@ -51,7 +51,7 @@ class InfoFilter(logging.Filter):
         return record.levelno == logging.INFO
 
 
-def setup_logger(*args, **kwargs):
+def setup_logger(verbose=False):
     """
     Setup and return the root logger object for the application.
     """
@@ -62,22 +62,22 @@ def setup_logger(*args, **kwargs):
     f = '%(asctime)s - %(name)s - %(levelname)s - %(message)s'
     formatter = logging.Formatter(f)
 
-    debug_handler = logging.StreamHandler(sys.stdout)
-    debug_handler.setLevel(logging.DEBUG)
-    debug_handler.addFilter(DebugFilter())
-    debug_handler.setFormatter(formatter)
+    if verbose:
+        debug_handler = logging.StreamHandler(sys.stdout)
+        debug_handler.setLevel(logging.DEBUG)
+        debug_handler.addFilter(DebugFilter())
+        debug_handler.setFormatter(formatter)
+        root.addHandler(debug_handler)
 
     info_handler = logging.StreamHandler(sys.stdout)
     info_handler.setLevel(logging.INFO)
     info_handler.addFilter(InfoFilter())
     info_handler.setFormatter(formatter)
+    root.addHandler(info_handler)
 
     error_handler = logging.StreamHandler(sys.stderr)
     error_handler.setLevel(logging.WARNING)
     error_handler.setFormatter(formatter)
-
-    root.addHandler(debug_handler)
-    root.addHandler(info_handler)
     root.addHandler(error_handler)
 
     return root